- when block_number and latest included, ignored `block_number` and returned latest instead. - did major clean-up and organization - rename module and plugin to BlockAccessServicePlugin, tried to keep consistency among all module and protobuf service definition
