+++ block-node/block-access/src/main/java/module-info.java
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
-import org.hiero.block.node.single.block.GetSingleBlockPlugin;
+import org.hiero.block.node.access.service.BlockAccessServicePlugin;
 
-module org.hiero.block.node.single.block {
+module org.hiero.block.node.access.service {
     uses com.swirlds.config.api.spi.ConfigurationBuilderFactory;
 
     requires transitive com.hedera.pbj.runtime;
@@ -10,5 +10,5 @@ module org.hiero.block.node.single.block {
     requires com.github.spotbugs.annotations;
 
     provides org.hiero.block.node.spi.BlockNodePlugin with
-            GetSingleBlockPlugin;
+            BlockAccessServicePlugin;
 }
